Rename: tempHeader -> Header
